Move to Projectile.TML
